Refactor, updates.
